Refactor utility imports to use the new helper utility file
- Updated imports in various modules to reference the new helper utility file instead of the previous utils.
- Added a new helper utility file that consolidates common utility functions such as ApiResponse, asyncForEach, isObject, exclude, pick, sanitize, and validateFields.
- Removed unused getFullName function from string utility file to streamline the codebase.
