+++ src/app/middlewares/pagination.middleware.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
 
-import { isObject } from '../../shared/utils';
+import { isObject } from '../../shared/utils/helper.util';
 import { InternalServerError } from '../../shared/utils/error.util';
 import { ObjectData } from '../../types';
 import { asyncHandler } from './async-handler.middleware';
