+++ src/app/modules/user/services/auth.service.ts
@@ -1,4 +1,4 @@
-import { exclude } from '../../../../shared/utils';
+import { exclude } from '../../../../shared/utils/helper.util';
 import { Repository } from '../../../../shared/utils/data-repo.util';
 import encrypterUtil from '../../../../shared/utils/encrypter.util';
 import { BadRequestError, UnauthorizedError } from '../../../../shared/utils/error.util';
